Changed how kick is handled
The original solution caused several problems where a MessageSender
thread might be in the middle of sending a message to a client at the
moment the socket is closed.

Instead the server sends an empty DataMessage to the client, and as soon
as it gets there the client closes its socket.