Too much stuff!
Refactored out userlist to its own panel. Did the same to the chatbox,
and implemented creation of groups (right-click selected users in
userlist).

Added a Group class model to help with the implementation of tabs.

I had to add a group variable to the privatemessage class so that when
the message gets sent back to the originating sender, the gui will know
which tab to put the message in.